:truck: Move `yaml_template` into `configuration` module
